Data_Pipeline/legacydb_to_bigquery_cloudfunc/requirements.txt
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.7
google-api-core==2.2.2
google-auth==2.3.3
google-cloud-bigquery==2.30.1
google-cloud-core==2.2.1
google-cloud-pubsub==2.9.0
google-crc32c==1.3.0
google-resumable-media==2.1.0
googleapis-common-protos==1.53.0
grpc-google-iam-v1==0.12.3
grpcio==1.42.0
grpcio-status==1.42.0
idna==3.3
libcst==0.3.23
mypy-extensions==0.4.3
numpy==1.21.4
packaging==21.3
pandas==1.3.4
proto-plus==1.19.8
protobuf==3.19.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyMySQL==1.0.2
pyparsing==3.0.6
python-dateutil==2.8.2
pytz==2021.3
PyYAML==6.0
requests==2.26.0
rsa==4.7.2
six==1.16.0
typing-inspect==0.7.1
typing_extensions==4.0.0
urllib3==1.26.7
